move givens_decomposition to _quadratic_hamiltonian.py
